Medical_app/requirements.txt
wygr==0.0.3
python-dotenv
langchain
langchain-openai 
openai
streamlit